employee-management-demo/backend/requirements.txt
sqlalchemy==2.0.23
fastapi==0.104.1
python-dotenv==1.1.1
uvicorn==0.34.3
email-validator==2.2.0
psycopg2-binary==2.9.10